replace 1900 files with copies of 1900
Something weird is happening with the 1900 met (no clue why), where
precip is orders of magnitude too low (1e-9 vs. 1e-5).  The changes
made here now replace that year (or facilitate similar replacements for
future years) with the following year.

This was done here so we don’t have to deal with trying to track down
vanishing files and trying to follow paper trails.